New-Fixture polished.
First I moved the New-Fixture, Create-File to separate file in
functions. When module is imported these functions are included
automatically.
I discarded Write-UsageForNewFixture. Usage can be looked up by
Get-Help, so help is the only one place to update if interface changes.
New-Fixture now outputs objects of the new files instead of writing to
Host. If file exists it writes to Warning.
Interface of the function is
kept the same, but when specifying only -Name parameter
fixture is created in the local folder.
Path accepts full, relative, UNC paths and "." for local directory.
Polished help.
And tweaked the default test as proposed in Issue 57.
New-Fixture tests added.