New-Fixture polished.
First I moved the New-Fixture, Create-File to separate file in
functions. When module is imported these functions are included
automatically.
I discarded Write-UsageForNewFixture. Usage can be looked up by
Get-Help.
New-Fixture now outputs objects of the new files instead of writing to
Host, if file exists it writes to Warning.  Interface of the function is
kept the same, but when specifying only -name parameter (by name)
fixture is created in the local folder. Path accepts full, relative, unc
paths and "." for local directory. Polished help. And tweaked the
default test as proposed in Issue 57.