TestDrive implemented
The test drive is created in describe and when moving to Context its
"state" is saved. When moving from the context files created in context
are deleted. This is done purely on full path basis so changes made in
context to Describe-scope files are kept. All test drive functions moved
to TestDrive. All tests regarding the test drive  moved to
TestDrive.ps1. Change to the Global scope polluting test were needed.
The test is not working properly anyway.  :(