It blocks are processed in memory.
The temp.ps1 script is not created for each It block, invoking
scriptblock inside an anonymous function is used instead.
Originally the $test was dumped into a Temp function, which was created
in a separate script and then dot-sourced to the main script, from which
it was invoked. That means the function was imported to the current
scope, so it translates exactly to what is used in this change set,
minus the overhead.
Also I cleaned up the mess in the It.Tests.