Mocking update (module scope separation)
Mocks in different modules now have completely separate records in $mockTable (denoted by a key of "Module||Command"), instead of sharing a Mock record, but having Blocks in separate scopes.
Code that operates on each Mock table record is now very nearly the same as it was before the Scope Isolation branch started, with a little bit of script block scope injection added.