It gets its own Pester scope
Mocks and Test Drive content created within "It" blocks are now cleared.

Mock call history used to reset back to zero at the end of any Context or Describe block; with the introduction of It scopes, call history is now only cleared for Mocks in the same scope.

Note:  One point here is still a bit awkward.  A mock's "scope" is set when the first mock for a particular command / module combination is created, and all calls to that module / command share a call history.
If the mock is overridden in a child pester scope, calls to that lower mock are still maintained in the call history.
We can probably change this so call history saves the Block's scope instead of the Mock's scope, which would correct the problem.