Test fix for direct execution
Mock tests had a problem if the test file was dot-sourced into the Global scope.
One of the dynamic test modules which is imported and later discarded contained an
exported function with the same name as one of the other functions in the test
script.  When the module is removed under these conditions, the original global
function is not restored.  This is not a defect in Pester; it's just how PowerShell
works.  Renaming that module function resolved the problem and allows the Mock
tests to execute, even when pressing F5 in the ISE (which is the equivalent
of dot-sourcing the test script.)