Mock call history change
Entries in $mockTable are now only cleared when there is no parent scope (which today, means when we exit a
Describe block.)  This way the call history of mocks defined in lower scopes are maintained, and can be checked
with Assert-MockCalled.

Assert-VerifiableMocks still requires the mock to be "live", though, since the Verifiable flag is associated with
the blocks that are cleared as we exit each Pester scope.