Changed default behavior of Assert-MockCalled
Command now identifies all calls to the mocked command by default (scope setting of "Describe").
If the test wants to narrow down the focus of the check, it can pass in -Scope Context or -Scope It.

This should result in fewer calls to Assert-MockCalled needing to specify a -Scope parameter (since
the most common use case should be to check for calls outside of the It block anyway), and will hopefully
also result in less impact to existing test scripts when this major version update is complete.