Adding Coverage Analysis (first pass)
Uses the AST to identify all commands in whatever script files the caller specifies, and sets empty PSBreakpoints on each of those commands.  After the tests finish, the breakpoints' HitCount properties are checked to determine coverage, and a report of any commands that have 0 HitCount is displayed.

Seems to work so far, but the tricky part is using this to analyze Pester's own test suite.  I've added several calls to Suspend-CoverageAnalysis and Resume-CoverageAnalysis (which basically correspond to Disable-PSBreakpoint and Enable-PSBreakpoint) to try to keep the number of false positives to a minimum, but I don't think we're ever going to be 100% accurate for our own internal tests.