NUnit OutputXML updated
The output XML used templates to output the nUnit xml report, that did not support using special
characters in Describe and It descriptions. I replaced it with XmlTextWriter that does all the
escaping automatically. Also the nUnit report was not compliant with the scheme. It now has single
test-suite root name by the Path (is there better option?). Scheme is not verified as well as the
special characters.

Fixing #122