Code coverage by function name
The hashtable that is passed to Invoke-Pester's -Coverage parameter can now have a key named Function (or f) to generate a coverage report for all commmands in that function.  The value assigned to this key may contain wildcards.

Some of the aliases for Path have been removed; now only Path and p are acceptable for that key.

The script doesn't currently generate an error if a hashtable contains both Function and StartLine and/or EndLine values, but if Function is present, StartLine / EndLine are ignored.  I should probably add either an error or warning output in this condition.