Remove tests testing usage of global scope
Removed the unused code that used to test the usage of global scope.
Looking through the code the only global variable is the global
TestDrive variable. This variable is provided for backwards
compatibility.

$pesterHome = Get-Module Pester | Select -ExpandProperty Path |
Split-Path
Get-ChildItem -Path $pesterHome -Include *.ps1, *.psm1 -Recurse |
Get-Content |
Select-String -Pattern '(\$global)|(-Scope\s+Global)'

Testing this properly needs running pester inside pester tests (not sure
if the whole test suite should be run or just a single test is enough)
and comparing the state of the global scope before and after run. Making
mental note of this and moving it to low priority folder.