Do not show warning if resolved code coverage path is folder
When wildcard in code coverage path is used it resolves and reports all
files that are not .ps1 or psm1 files. This included folder paths which
might appear to the user that no script in the path is  included, which
is not true. Warnings for folder paths are no longer shown.