Common parent path based on all coverage files, not just missed
If you're analyzing coverage on, for example:

    Folder\Subfolder1\File.ps1
    Folder\Subfolder2\File.ps1

Where the file names are not unique, but the paths are, in the previous commit you'd just see 'File.ps1' if only one of the files contained missed commands.  In this commit, you'll see 'Subfolder1\File.ps1'.  The "common parent folder" check is done against the whole list of analyzed files, rather than just the files that happened to have missed commands.