Initial implementation of BeforeEach / AfterEach
User-Defined setup and teardown blocks that execute before and after each "It" block.  Scope rules based on what I read in rspec's documentation; Setup blocks are executed from the Describe scope first, then Context.  For teardown, it's in reverse; the innermost (Context) teardown blocks are executed, then the outer (Describe) blocks.

For now, the code allows you to define multiple blocks in the same scope, and they are executed in the order they were defined.

BeforeEach and AfterEach blocks can be defined anywhere within a Context or Describe block; they are parsed and loaded before the body of the Context / Describe is executed.  This is different from how PowerShell code normally behaves, but was a desired feature in the initial discussion.