Assert-MockCalled fix
The value of the ModuleName parameter in Assert-MockCalled now defaults to Pester's current session state.  Normally, this won't change its behavior, but inside an InModuleScope block, you no longer need to explicitly specify -ModuleName when calling Assert-MockCalled for mocks in the same module.  (This brings Assert-MockCalled functionality back in line with the Mock command, which also doesn't require you to specify -ModuleName inside an InModuleScope block.)