Errors reported from Describe and Context blocks
Quick update to make terminating errors that occur in a Describe or Context block cause a failed "test" to appear in the output.

Because we're just piggybacking onto the existing PesterState.AddTestResult and Write-PesterResult code, the output looks as though it came from an It statement, though the name is always "Error occurred in Context block" or "Error occurred in Describe block".  If this style of output is undesirable, a more extensive code update would be required.