Pester.bat no longer enables strict mode.
This change was originally made to ensure that Pester's unit tests are run with strict mode enabled, so we would catch any bugs.  However, making this change in Pester.bat affects all users of Pester (or at least the ones who use the convenience batch file for integration.)

Pester's unit test scripts have since each been modified to enable Strict Mode at the beginning of each file, and we should no longer require this to be set in Pester.bat.  This way, users of Pester are not forced to write their test scripts in a way that is StrictMode-compatible.