Describe and Context fix
Fixture parameters to Describe was mistakenly marked as Mandatory in a recent commit.  Instead of that, its previous behavior was to be optional, but to throw a more user-friendly error message if it's not specified.

Added tests to make sure this doesn't happen again, and also applied the same logic / tests to the Context command.