"Should Be" and "Should BeExactly" updates
Original code was behaving a little weirdly, and caused some tests to fail in conjunction with the new Should on Collections behavior.  This test used to pass:

1 | Should Be @(1,2,3,4,5)

And this test used to fail:

@(1,2,3,4,5) | Should Be @(1,2,3,4,5)

This situation has been corrected in both cases.  "Should Be" now compares the elements of arrays on both sides (in order) and only returns true if all of the elements match with the -eq operator (or -ceq, in the case of BeExactly).