More Should Be and BeExactly updates
Moved these assertions into the same file, which makes sense for organization when they use the same helper functions.

Be / BeExactly now make use of the Compare-Object cmdlet, which has demonstrated more consistent behavior when arrays containing other arrays are passed to Should (though I can't imagine users actually need to do that, at least we handle it properly.)  It may also be easier to explain the behavior of "Should Be" in terms of it being a wrapper around Compare-Object.

There were a couple of tests which relied on the behavior of the eq operator in the old implementation of Should Be (automatic casting of operands).  These tests have been updated to use Should Not Be $null instead of Should Be $true (where a non-boolean value is piped in.)