Code called inside InModuleScope now retains file / line information
This causes failed tests to output the proper file and line from the tests.ps1 file, and also allows for setting breakpoints inside these blocks.

There are two other places that currently use Set-ScriptBlockScope and may lose similar information:  Mocks, and BeforeEach / AfterEach.  I'm not sure we can do much about the latter, but we can possibly add this same type of change to the Mocking code, if needed.