Reporting correct file name
For exceptions that are generated by something other than the Should command, test failure output now contains the file path and line number of whatever code produced the exception, instead of mixing the file path of the Tests.ps1 script and the line number from whatever file contained the error-producing code.

May revisit this to either change it to report the path / line number of the Tests.ps1 file (as failed Should assertions currently do), or output both sets of information.