Should be finished
Operators can now signal to Should that they operate on the entire input set in one go by defining a "Pester<Operator>AcceptsArrayInput" function which returns $true.  Be / BeExactly / BeNullOrEmpty take advantage of this new behavior, and the other operators now behave exactly as they did before, with the option of changing them later if it makes sense.