Add verbose output for string comparison for Be and BeExactly
To make subtle differences in strings easier to spot, a VS inspired
output is used to point out the exact difference.

Adding this for Be and BeExactly.
Describing Special output for string in Should Be
 [-] outputs verbose message 165ms
   String lengths are both 3. Strings differ at index 2.
   Expected: {abd}
   But was:  {abc}
   -------------^
   at line: 3 in C:\temp\VerboseOutputExamples.ps1

Move the definition and tests of BeExactly to the Be and Be tests to
share the functionality among the assertions.

Describing Special output for string in Should BeExactly
 [-] outputs verbose message 12ms
   String lengths are both 1. Strings differ at index 0.
   Expected: {a}
   But was:  {A}
   -----------^
   at line: 7 in C:\temp\VerboseOutputExamples.ps1