Still more DynamicParam fixes
To retrieve the dynamic parameters for a function, the code now executes a script block which is bound to the scope of the same module as the mocked function's original location.  This allows the dynamic parameters to be fetched successfully even when they reference script-scoped variables, and the caller is in a different module scope.