Refactoring to allow more thorough testing of It command
For now, this is accomplished by moving the entire implementation of It to another function, and giving that function a parameter named $Pester.  Eventually, a cleaner approach will be to pass around a PesterState object as a parameter throughout the entire internal implementation of the code, but right now, too much of the code relies on being able to dynamically resolve a variable named $Pester, and it'll be easier to fix that situation from the bottom up.