Added new parameter set for specifying export path / format
Will allow us to define other types of exports later on.  For now, the only option for format is LegacyNUnitXml, which matches the behavior of the existing -OutputXml parameter.

Next step is to tweak the NUnit format with a new, more compatible version (leaving the current output of LegacyNUnitXml alone for as long as we choose to support it.)