Add Result property to TestResult
TestResult objects desperately need more states than just passed, failed
(or True/False) as it is now represented. To be able to explicitly show
the test was skipped, or is pending (inconclusive/not implemented). For
that reason I am adding Result property to the status object that
further specifies the Passed (bool) property.