Add Skip and Pending keywords
Add Skip and Pending keywords that use special exceptions to jump out of
It block. Ideally these keywords should be placed as first in It
ScriptBlock. But there is no way I can enforce it, or detect them
without using AST and breaking PowerShell 2 compatibility.

If we dropped the PowerShell 2 support, the AST could be used to detect
the Skip or Pending functions and skip the It without throwing any
special exception (simply by implementing the check inside It and not
running the test body if there is Skip/Pending present).