Changelog and description
I added Result property to TestResult objects and kept the Passed
boolean property. The Result can have four values: Failed, Skipped,
Pending, Passed, ordered from the most important to the least important.

Both Pending and Skipped can be forced by setting -Pending (-Skip)
parameters on It. The Pending is also used when It's ScriptBlock is
empty or contains only comments.

(The New-Fixture should be edit in some way to take advantage of this,
not sure what to put there ATM)

The new types of Results are taken into account when exporting the
result into nUnit report.

The new output is now gray, because that is the only usable color on
dark and light background that was not used yet. I will probably add
Light and Dark color sets and chceck the currenct background color in
some way. But I don't want to mix this change with the current change.

I also added Strict property to every TestResult and to PesterState
object. Strict parameter is available on Invoke-Pester to force anything
than Passed to Fail.

That is all I guess. Everything should be visible in the commit history
(I will squash it for the final merge).