Fix the ItImpl parameters
Fixed the ItImpl parameters to make it work correctly with the Skipped
and Pending. I also set the ScriptBlock to empty scriptblock after
validating all parameters. That way the Invoke-Test function interface
is kept strict (it requires a ScriptBlock) while the desired
functionality is kept.  I am not sure if this is hundred percent
correct, because I am not sure where the boundary between Pester and
'Common Testing Framework' should be and what these should and should
not implement.