Add basic tests that validate the manifest, changelog and tags
When a new version of Pester is released there are three things that need to be edited and done:
- the module manifest must be updated to reflect the correct version
- the changelog must be updated to contain the current version and release change notes
- the commit must me tagged with the correct version

I had problems remembering all these so there are few basic tests that validate the build.

There are few tests that run only when the git.exe is available. So after tagging a new version you should run the .\bin\Pester.bat from the git to validate the changes. I will have to check on TC if all the tests run.

Known bug: this format you might use with tags 3.0.6*-beta* is not supported (yet?)

Function that ups the version automatically might follow.