Performance improvements
I did some profiling of the module today, and identified several hotspots that have been improved in this commit.  The main three are:

ScriptProperties on the PesterState object (FailedCount, etc) have been replaced with normal properties that are incremented in the AddTestResult function.

Parsing performed for the BeforeEach / AfterEach commands is run a lot and was expensive; this code has been moved into some compiled C# for more speed.

Calls to Get-IgnoreErrorPreference were also adding up quite a bit.  These have been replaced with a lookup of a script-scope variable instead.  This results in a tiny bit of duplication in the Mock code, but it's well worth it.

There are a few lesser changes as well.  On my WMF 5.0 system, the time it takes to execute Pester's own test suite has gone down by around one third.