Making Pester StrictMode compliant again.
We received a bug report that Pester was breaking if StrictMode were set for the whole PowerShell session.  (We've been enabling it at the beginning of each test script, but apparently that's not enough to ensure compliance.  We'll need to enable it before calling Invoke-Pester in our build jobs instead.)

All of these errors have been fixed.