Build script update
nuget apparently has a bug when adding files that have no extension, and the recent addition of LICENSE was causing the pack step to fail.  There's a workaround for this to add a * character to the end of the file name, which appears to work.

I've also tweaked the psake build script slightly to make sure the unversion step works regardless of the $currentDir.