Added new test, cleaned up some code
Several references to $Exception were being made when the object is actually an ErrorRecord.  I've renamed those bits to improve clarity, and also added type information to a param block for the same reason.

Added new test to make sure that Get-PesterResult isn't monkeying with failure messages, as mentioned in #282 .