Initial fix for #286
This works, but there might be some weirdness that can still happen if you mock a global function in conjunction with -ModuleName or InModuleScope.  (The global function gets renamed for the duration of the mock, but the mock is only available from that single module.  You can't mock the same global function from two different modules at once.)

Not sure if it's worth worrying about that edge case, since this "mocking global functions" scenario  is already pretty rare to begin with.