Workaround for GetDynamicParameters() bug
Weird ActiveDirectory module is weird.

Fixes #294.  There are no automated tests to verify this yet; we either need to run the test on a 2008 R2 system with the RSAT feature installed, or write some C# code to reproduce the behavior of the AD module.  (The latter would be preferable, but will take some more time.)