Fixes for #315 and #269
The new parameter for Assert-MockCalled is called `-NotOtherwise` for now, but I think we may want to come up with a better name.  Either way, the functionality is implemented now.

Assert-MockCalled now throws the same type of ErrorRecord objects as the Should command, which gets us better output as noted in #269.