Fixes for #329 and #319, and bug fix in Should Be
Figured out how to make this work on v3+.  It required a little bit more futzing around with Reflection and internals, but what can ya do.

Also fixed a bug in Should Be: several string parameters were marked as mandatory but did not allow an empty string for the input that was piped to Should.