Potential fix for #345
This is an ugly hack, but it does appear to work so far.  First, we strip out all information about parameter sets from the param block that contains the copy of the original function's dynamicparam code.  This way it should be safe to call without throwing exceptions due to missing mandatory parameters or ambiguous parameter sets.

However, this breaks the tests that depend on being able to use $PSCmdlet.ParameterSetName.  So, we add an extra parameter (named ${P S Cmdlet}, which is virtually guaranteed to never collide with an actual parameter name) into the script block, and pass in a copy of $PSCmdlet from out in the mock's bootstrap function.  Since this script block contains an unmodified copy of the mocked function's param block, all information about parameter sets is intact.
This new parameter is written over top of the automatic $PSCmdlet variable just before the copy of the dynamicparam block's statements begins to execute.

If it becomes necessary, we can do the same thing for other automatic variables this way as well.