Initial commit
There's a crapload of code cleanup to be done, but the basic functionality for Should -Be and Should -BeExactly is here.  Other operators still need some code updates to work with the new system, but I'm going to hold off on that until I clean things up around just -Be.