Added failing tests.
Two bugs here.  One, if a mock exists but no blocks match their parameter filter, the original command is called from Pester's session state instead of from the session state of the original caller of the mock.  This can sometimes cause bugs in people's scripts.

Second, inside of Mock / Validate-Command, calls are being made to Get-Content in such a way that if the test author mocks it, Pester's internals wind up triggering the mock (because it runs code in their module scope.)  This can also lead to errors.