Bug fixes
There's still a potential gotcha here if a non-advanced function is mocked, and there are no matching parameter filters.  In that case, there's no $PSCmdlet variable in the mock bootstrap function, and I don't know how else to find the caller's SessionState object.  In that circumstance, I opted to execute the original command from the scope where the mock is defined, which should be correct most of the time.