Refactoring assertion operators
The test method for each assertion is now responsible for both returning a success flag, and a failure message if the success flag is false.  This way, each assertion's functions can contain custom logic based on whatever parameters they choose to define in the new dynamic parameter sets, and we don't have to try to have the Should command figure out how to call the "FailureMessage" functions instead.

Still need to add tests for the new syntax of other operators.