Make Pester run on Nano server
There were two changes that needed to be made to make Pester work on
Nano server. First I had to conditionally import Get-CimSession instead
of Get-WmiObject that is not available on nano server.  The other thing
that I replaced was the BraceFinder that was written and compiled via
Add-Type. The PowerShell implementation is about 10 times slower (but
it's 1 vs 10 milliseconds).

The rest of the fixes needed to be done in test files.
1) DirectoryInfo serializes to just Name instead of FullName which
caused the most of the failed tests, so I had to resort to storing the
full name in the Global:TestDrive variable instead of the DirectoryInfo
object. This is a breaking change, but I don't see a better option.

2) I had to conditionally skip some of the xml tests because the
required properties and methods to handle schema validation are not
available on .net core.