Updates to better support Pester on Nano
Update Coverage tests to use a different resource. The WindowsFeature resource
doesn't exist on Nano, but the Environment resource does
Update the TestResult tests to only validate against schema when the Validate
method is available on the system. Nano, for example, doesn't have this method
available because of the differences between Core and Full CLR