Initial updates for arbitrary group nesting
Working on #554 .  Still a ways to go, but at least Mocking and Setup/Teardown work with the new plumbing.

TestResult property still exists on $pester for backward compatibility (which is also why the XML code still works as-is for the moment), but the Describe / Context names are only provided on a best-effort basis.  Anyone using the new ability to nest Describe / Context any way they like should be using the new TestActions hierarchy instead.