Proxy cmdlet support for safe cmdlet lookup
The pester module will fail to load when someone has a module already loaded that defines a cmdlet with the same name as one of the safe commands. An example of this is FormatPx, which defines Format-Table and Out-String proxy cmdlets (not functions, cmdlets). I have FormatPx loaded in my profile, and when I try to load Pester or invoke a Pester command using 5.1.14393, it fails. This change fixes the problem by adding the -All parameter to all Get-Command calls used to build up the list of safe cmdlets.