Make Should Throw case insensitive
I replaced the original Contains with IndexOf and a $null check, becuase
Contains returns $True for $null input. The $null in this context means
that users did not specify any expected message and hence all messages
should pass the Get-MessagesMatch filter.

There was also a test case which made sure that empty string expectation
returns $False. I removed that case, because it complicates the
contract, and no one is likely using it as it would always fail the
assertion. Instead replaced it with cases that ensure that empty-string
error message will match empty-string expectation.

I also removed the $ExpectedErrorId parameter from Get-DoMessagesMatch,
because it was not used. I guess the idea was to implement the logic of
"id and message must both match" inside of the Get-DoMessageMatch
function, but it is implemented elsewhere.

This is a breaking change, and so is making the comparison
case-insensitive.

Fixes #577