Add option to Invoke-Pester to filter output
Adds Show parameter to Invoke-Pester that filters the output of Pester.

The parameter takes flags of a new type [Pester.OutputTypes]
with the following options: Default, Passed, Failed, Pending,
Skipped, Inconclusive, Describe, Context, Summary.

And three special options: None, All and Fails. The All option is
the default and prints all the output. The None option is replacement
for the -Quiet parameter (still present and not deprecated).
The Fails option shows everything except Passed.

The options can be arbitrarily mixed, for example to provide a very
verbose ouptut that includes only failed tests and summary use:
Invoke-Pester -Show Summary, Failed

The Summary, Failed option is deliberaterly not chosen as the Fails
option because I assume that the Describe and Context information
are valuable for the general user, and it's much harder to spell out
the definition of Fails than the definition of Summary, Failed.

The Summary, Failed option can be added to the OutputTypes as a
special case but I could not come up with a good name. Feel free to
suggest it.

Another option that I considered was skipping empty describes
and Context, (for example by posponing descibe till a test result is
written), but that would change the behavior for tests with long Before*
blocks, as well as the general behavior for empty Describes. And I don't
want to put much logic in writing output to keep it straightforward
for the users.

Fix #645
Fix #568